- Move common enums to a separate file. - Add Format and Encoding enums.
