+++ src/common/dpba_interface.h
@@ -6,41 +6,13 @@
 #include <span>
 #include <string>
 #include <utility>
+#include "enums.h"
 
 #ifndef DBPS_EXPORT
 #define DBPS_EXPORT
 #endif
 
 namespace dbps::external {
-
-// Captures the data type of the data batch elements.
-// Intentionally similar to parquet::Type to ease mapping and for compatibility with a known enum.
-struct Type {
-    enum type {
-        BOOLEAN = 0,
-        INT32 = 1,
-        INT64 = 2,
-        INT96 = 3,
-        FLOAT = 4,
-        DOUBLE = 5,
-        BYTE_ARRAY = 6,
-        FIXED_LEN_BYTE_ARRAY = 7
-    };
-};
-
-// Intentionally similar to arrow::CompressionCodec
-struct CompressionCodec {
-    enum type {
-        UNCOMPRESSED = 0,
-        SNAPPY = 1,
-        GZIP = 2,
-        LZO = 3,
-        BROTLI = 4,
-        LZ4 = 5,
-        ZSTD = 6,
-        LZ4_RAW = 7
-    };
-};
 
 // ++++ To docs:
 // - While handle to EncryptionResult exists, ciphertext() is guaranteed to return the result.
